Move voice keying related code to audio.c
