+++ src/stoptx.c
@@ -21,7 +21,7 @@
 *
 *--------------------------------------------------------------*/
 
-
+#include "audio.h"
 #include "clear_display.h"
 #include "err_utils.h"
 #include "globalvars.h"
@@ -48,7 +48,7 @@ int stoptx(void) {
 	    }
 	}
     } else if (trxmode == SSBMODE) {
-	IGNORE(system("pkill -SIGTERM -n play_vk"));
+	stop_vk();
 	return 0;
     } else {
 	return 1;
